Extracting elementsEqual matcher to ElementsEqual.swift
