+++ Tests/NimbleTests/Matchers/ElementsEqualTests.swift
@@ -1,0 +1,37 @@
+//
+//  ElementsEqualTests.swift
+//  Nimble
+//
+//  Created by Luciano Almeida on 7/1/18.
+//  Copyright © 2018 Jeff Hui. All rights reserved.
+//
+
+import Foundation
+
+import Foundation
+import XCTest
+import Nimble
+
+final class ElementsEqualTest: XCTestCase, XCTestCaseProvider {
+
+    func testSequenceElementsEqual() {
+        failsWithErrorMessageForNil("expected to elementsEqual <nil>, got <nil>") {
+            expect(nil as [Int]?).to(elementsEqual(nil as [Int]?))
+        }
+        let sequence = [1, 2]
+        failsWithErrorMessageForNil("expected to elementsEqual <[1, 2]>, got <nil>") {
+            expect(nil as [Int]?).to(elementsEqual(sequence))
+        }
+
+        failsWithErrorMessageForNil("expected to elementsEqual <nil>, got <[1, 2]>") {
+            expect(sequence).to(elementsEqual(nil as [Int]?))
+        }
+
+        let sequence1 = [1, 2, 3]
+        let sequence2 = [1, 2, 3, 4, 5]
+        expect(sequence1).toNot(elementsEqual(sequence2))
+        expect(sequence1).toNot(elementsEqual([3, 2, 1]))
+        expect(sequence1).to(elementsEqual([1, 2, 3]))
+
+    }
+}
